@@ -1,11 +1,11 @@
 name: flutter_reaction_button
 description: Flutter button reaction it is fully customizable widget such as Facebook reaction button
-version: 1.0.7+3
+version: 1.0.8
 homepage: https://github.com/GeekAbdelouahed/flutter-reaction-button
 repository: https://github.com/GeekAbdelouahed/flutter-reaction-button.git
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
