@@ -1,9 +1,6 @@
-import 'package:avataaar_image/avataaar_image.dart';
-
 class Comment {
-  final Avataaar avatar;
   final String name;
   final String content;
 
-  const Comment({this.avatar, this.name, this.content});
+  const Comment({required this.name, required this.content});
 }
